backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.29,<1.0
aiosqlite>=0.20,<1.0
pydantic-settings>=2.2,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
python-multipart>=0.0.9,<1.0
croniter>=6.0,<7.0
psutil>=6.0,<7.0
